# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Terminal AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SshConnectionAuthType =
  (typeof SshConnectionAuthType)[keyof typeof SshConnectionAuthType];

export const SshConnectionAuthType = {
  password: "password",
  key: "key",
} as const;

export interface SshConnection {
  id: number;
  name: string;
  host: string;
  port: number;
  username: string;
  authType: SshConnectionAuthType;
  createdAt: string;
  updatedAt: string;
}

export type CreateSshConnectionBodyAuthType =
  (typeof CreateSshConnectionBodyAuthType)[keyof typeof CreateSshConnectionBodyAuthType];

export const CreateSshConnectionBodyAuthType = {
  password: "password",
  key: "key",
} as const;

export interface CreateSshConnectionBody {
  name: string;
  host: string;
  port: number;
  username: string;
  authType: CreateSshConnectionBodyAuthType;
  password?: string;
  privateKey?: string;
  passphrase?: string;
}

export interface Snippet {
  id: number;
  title: string;
  command: string;
  description?: string;
  category?: string;
  createdAt: string;
}

export interface CreateSnippetBody {
  title: string;
  command: string;
  description?: string;
  category?: string;
}

export interface AiSettings {
  id: number;
  /** Masked API key (only last 4 chars shown) */
  apiKey?: string;
  endpointUrl: string;
  modelName: string;
  hasApiKey: boolean;
  systemPrompt?: string;
  /** Whether a GitHub personal access token is stored */
  hasGithubToken: boolean;
  /** GitHub repo in owner/repo format */
  githubRepo?: string;
}

export interface UpdateAiSettingsBody {
  apiKey?: string;
  endpointUrl?: string;
  modelName?: string;
  systemPrompt?: string;
  /** GitHub personal access token (leave blank to keep current) */
  githubToken?: string;
  /** GitHub repo in owner/repo format (e.g. myuser/ai-memory) */
  githubRepo?: string;
}

export interface SendChatMessageBody {
  message: string;
  /** Recent terminal output for AI context */
  terminalContext?: string;
  /** Whether this message is part of an autonomous agent loop */
  agentMode?: boolean;
  /** The original task assigned to the agent (for context) */
  agentTask?: string;
  /** Current step number in the agent loop */
  agentStep?: number;
  /** If true, always save this user message to the memory file */
  forceMemory?: boolean;
}

export type ChatResponseRole =
  (typeof ChatResponseRole)[keyof typeof ChatResponseRole];

export const ChatResponseRole = {
  assistant: "assistant",
} as const;

export interface ChatResponse {
  message: string;
  /** Command the AI wants to execute (if any) */
  command?: string;
  /** Agent signals task is complete (agent mode only) */
  done?: boolean;
  role: ChatResponseRole;
  /** Whether the AI wrote something to memory this turn */
  memorySaved?: boolean;
  /** Whether this turn triggered a GitHub sync */
  synced?: boolean;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  command?: string;
  terminalContext?: string;
  createdAt: string;
}