# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Terminal AI API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all SSH connections
 */
export const ListSshConnectionsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  host: zod.string(),
  port: zod.number(),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListSshConnectionsResponse = zod.array(
  ListSshConnectionsResponseItem,
);

/**
 * @summary Create a new SSH connection
 */
export const createSshConnectionBodyPortDefault = 22;

export const CreateSshConnectionBody = zod.object({
  name: zod.string(),
  host: zod.string(),
  port: zod.number().default(createSshConnectionBodyPortDefault),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  password: zod.string().optional(),
  privateKey: zod.string().optional(),
  passphrase: zod.string().optional(),
});

/**
 * @summary Get a single SSH connection
 */
export const GetSshConnectionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSshConnectionResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  host: zod.string(),
  port: zod.number(),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update an SSH connection
 */
export const UpdateSshConnectionParams = zod.object({
  id: zod.coerce.number(),
});

export const updateSshConnectionBodyPortDefault = 22;

export const UpdateSshConnectionBody = zod.object({
  name: zod.string(),
  host: zod.string(),
  port: zod.number().default(updateSshConnectionBodyPortDefault),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  password: zod.string().optional(),
  privateKey: zod.string().optional(),
  passphrase: zod.string().optional(),
});

export const UpdateSshConnectionResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  host: zod.string(),
  port: zod.number(),
  username: zod.string(),
  authType: zod.enum(["password", "key"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete an SSH connection
 */
export const DeleteSshConnectionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all command snippets
 */
export const ListSnippetsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  command: zod.string(),
  description: zod.string().optional(),
  category: zod.string().optional(),
  createdAt: zod.coerce.date(),
});
export const ListSnippetsResponse = zod.array(ListSnippetsResponseItem);

/**
 * @summary Create a new command snippet
 */
export const CreateSnippetBody = zod.object({
  title: zod.string(),
  command: zod.string(),
  description: zod.string().optional(),
  category: zod.string().optional(),
});

/**
 * @summary Update a command snippet
 */
export const UpdateSnippetParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSnippetBody = zod.object({
  title: zod.string(),
  command: zod.string(),
  description: zod.string().optional(),
  category: zod.string().optional(),
});

export const UpdateSnippetResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  command: zod.string(),
  description: zod.string().optional(),
  category: zod.string().optional(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a command snippet
 */
export const DeleteSnippetParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get AI provider settings
 */
export const GetAiSettingsResponse = zod.object({
  id: zod.number(),
  apiKey: zod
    .string()
    .optional()
    .describe("Masked API key (only last 4 chars shown)"),
  endpointUrl: zod.string(),
  modelName: zod.string(),
  hasApiKey: zod.boolean(),
  systemPrompt: zod.string().optional(),
  hasGithubToken: zod
    .boolean()
    .describe("Whether a GitHub personal access token is stored"),
  githubRepo: zod
    .string()
    .optional()
    .describe("GitHub repo in owner\/repo format"),
});

/**
 * @summary Update AI provider settings
 */
export const UpdateAiSettingsBody = zod.object({
  apiKey: zod.string().optional(),
  endpointUrl: zod.string().optional(),
  modelName: zod.string().optional(),
  systemPrompt: zod.string().optional(),
  githubToken: zod
    .string()
    .optional()
    .describe("GitHub personal access token (leave blank to keep current)"),
  githubRepo: zod
    .string()
    .optional()
    .describe("GitHub repo in owner\/repo format (e.g. myuser\/ai-memory)"),
});

export const UpdateAiSettingsResponse = zod.object({
  id: zod.number(),
  apiKey: zod
    .string()
    .optional()
    .describe("Masked API key (only last 4 chars shown)"),
  endpointUrl: zod.string(),
  modelName: zod.string(),
  hasApiKey: zod.boolean(),
  systemPrompt: zod.string().optional(),
  hasGithubToken: zod
    .boolean()
    .describe("Whether a GitHub personal access token is stored"),
  githubRepo: zod
    .string()
    .optional()
    .describe("GitHub repo in owner\/repo format"),
});

/**
 * @summary Send a message to the AI with optional terminal context
 */
export const SendChatMessageBody = zod.object({
  message: zod.string(),
  terminalContext: zod
    .string()
    .optional()
    .describe("Recent terminal output for AI context"),
  agentMode: zod
    .boolean()
    .optional()
    .describe("Whether this message is part of an autonomous agent loop"),
  agentTask: zod
    .string()
    .optional()
    .describe("The original task assigned to the agent (for context)"),
  agentStep: zod
    .number()
    .optional()
    .describe("Current step number in the agent loop"),
  forceMemory: zod
    .boolean()
    .optional()
    .describe("If true, always save this user message to the memory file"),
});

export const SendChatMessageResponse = zod.object({
  message: zod.string(),
  command: zod
    .string()
    .optional()
    .describe("Command the AI wants to execute (if any)"),
  done: zod
    .boolean()
    .optional()
    .describe("Agent signals task is complete (agent mode only)"),
  role: zod.enum(["assistant"]),
  memorySaved: zod
    .boolean()
    .optional()
    .describe("Whether the AI wrote something to memory this turn"),
  synced: zod
    .boolean()
    .optional()
    .describe("Whether this turn triggered a GitHub sync"),
});

/**
 * @summary Get recent chat history
 */
export const GetChatHistoryResponseItem = zod.object({
  id: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  command: zod.string().optional(),
  terminalContext: zod.string().optional(),
  createdAt: zod.coerce.date(),
});
export const GetChatHistoryResponse = zod.array(GetChatHistoryResponseItem);